% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Championnat_Constr_F1}
\alias{Championnat_Constr_F1}
\title{Datasets about constructor championship in F1}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Championnat_Constr_F1)
}
\description{
Datasets about constructor championship in F1
}
\keyword{datasets}
