% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Championnat_Pilote_F1}
\alias{Championnat_Pilote_F1}
\title{Dataset about Championship F1}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Championnat_Pilote_F1)
}
\description{
Dataset about Championship F1
}
\keyword{datasets}
