% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Classements_Constructeurs_Finaux}
\alias{Classements_Constructeurs_Finaux}
\title{Dataset about Final constructor Championship}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Classements_Constructeurs_Finaux)
}
\description{
Dataset about Final constructor Championship
}
\keyword{datasets}
