% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Classements_Pilotes_Finaux}
\alias{Classements_Pilotes_Finaux}
\title{Dataset about F1}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Classements_Pilotes_Finaux)
}
\description{
Dataset about F1
}
\keyword{datasets}
