% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions2.R
\name{compare2}
\alias{compare2}
\title{Title}
\usage{
compare2(pilote1 = "L. HAMILTON", pilote2 = "N. ROSBERG")
}
\arguments{
\item{pilote1}{pilote de f1}

\item{pilote2}{second pilote de f1}
}
\value{
comparaison entre les pilotes, plusieurs barplots concernant leurs statistiques personnelles
}
\description{
Title
}
