% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Courses_F1}
\alias{Courses_F1}
\title{Dataset about KiteSurf}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Courses_F1)
}
\description{
Dataset about KiteSurf
}
\keyword{datasets}
