% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_Pilotes}
\alias{df_Pilotes}
\title{Dataset about Karting}
\format{
An object of class \code{data.frame} with 337 rows and 6 columns.
}
\usage{
data(Comparaison_Pilotes)
}
\description{
Dataset about Karting
}
\keyword{datasets}
