% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_extraction.R
\name{find_circuits_in_urls_vec}
\alias{find_circuits_in_urls_vec}
\title{Vectorised version of .}
\usage{
find_circuits_in_urls_vec(urls)
}
\arguments{
\item{urls}{Character. url's}
}
\value{

}
\description{
Vectorised version of .
}
