% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_extraction.R
\name{find_qualifying_urls}
\alias{find_qualifying_urls}
\title{Scrape all urls that link to qualifying statistics.}
\usage{
find_qualifying_urls(main_url, base_url = "https://www.formula1.com")
}
\arguments{
\item{main_url}{url from formula1.com ending with '/qualifying.html'}

\item{base_url}{The base url. Default 'https://www.formula1.com'.}
}
\value{
character with urls.
}
\description{
Scrape all urls that link to qualifying statistics.
}
\examples{
url <- 'https://www.formula1.com/en/results.html/2022/races/1114/great-britain/qualifying.html'
quali_urls <- find_qualifying_urls(url)
print(quali_urls)
}
