% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{get_capital_name}
\alias{get_capital_name}
\title{Extract the year from a url like 'https://www.formula1.com/en/results.html/2022/races/'}
\usage{
get_capital_name(driver_name)
}
\arguments{
\item{driver_name}{Character.}
}
\value{

}
\description{
Extract the year from a url like 'https://www.formula1.com/en/results.html/2022/races/'
}
