% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualification_data.R
\name{get_qualification_stats}
\alias{get_qualification_stats}
\title{Scrape the qualification result statistics from a given year.}
\usage{
get_qualification_stats(year)
}
\arguments{
\item{year}{int.}
}
\value{

}
\description{
Scrape the qualification result statistics from a given year.
}
\examples{
# get_qualification_stats(2021)
}
