% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_extraction.R
\name{get_qualification_urls}
\alias{get_qualification_urls}
\title{Get the url with qualification results from a year.}
\usage{
get_qualification_urls(year)
}
\arguments{
\item{year}{int. Year that you want to scrape.}
}
\value{

}
\description{
Get the url with qualification results from a year.
}
