% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_extraction.R
\name{get_race_results_url}
\alias{get_race_results_url}
\title{Given a year find the race resuslts.html}
\usage{
get_race_results_url(year)
}
\arguments{
\item{year}{int. The year you want to scrape}
}
\value{
Character. A url.
}
\description{
Given a year find the race resuslts.html
}
\examples{
# get_race_results_url(2019)
}
