% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{get_year_from_url}
\alias{get_year_from_url}
\title{Extract the year from a url like 'https://www.formula1.com/en/results.html/2022/races/'}
\usage{
get_year_from_url(url)
}
\arguments{
\item{url}{Character.}
}
\value{
Character. Year.
}
\description{
Extract the year from a url like 'https://www.formula1.com/en/results.html/2022/races/'
}
