% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liste_Pilotes}
\alias{liste_Pilotes}
\title{Dataset about Rodeo}
\format{
An object of class \code{character} of length 370.
}
\usage{
data(liste_Pilotes)
}
\description{
Dataset about Rodeo
}
\keyword{datasets}
