% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Noms_Epreuves}
\alias{Noms_Epreuves}
\title{Dataset about handball}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Noms_Epreuves)
}
\description{
Dataset about handball
}
\keyword{datasets}
