% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Qualifs_F1}
\alias{Qualifs_F1}
\title{Dataset about things}
\format{
An object of class \code{list} of length 71.
}
\usage{
data(Qualifs_F1)
}
\description{
Dataset about things
}
\keyword{datasets}
