% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dataset.R
\name{rem_empty_elements}
\alias{rem_empty_elements}
\title{Remove all empty elements from a list.}
\usage{
rem_empty_elements(l)
}
\arguments{
\item{l}{List}
}
\value{

}
\description{
Remove all empty elements from a list.
}
