% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{scrape_qualification}
\alias{scrape_qualification}
\title{Extract the table with qualifying stats from the website.}
\usage{
scrape_qualification(url)
}
\arguments{
\item{url}{Character. URL from formula1.com that ends with
qualifying.html'. It is used to scrape}
}
\value{

}
\description{
Extract the table with qualifying stats from the website.
}
