% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{scrape_qualifying_stats}
\alias{scrape_qualifying_stats}
\title{Vectorised version to scrape multiple urls at once.}
\usage{
scrape_qualifying_stats(qualifying_urls)
}
\arguments{
\item{qualifying_urls}{Character. URL's scraped with
the fun \code{find_qualifying_urls()}}
}
\value{

}
\description{
Vectorised version to scrape multiple urls at once.
}
\examples{
url <- 'https://www.formula1.com/en/results.html/2022/races/1114/great-britain/qualifying.html'
quali_urls <- find_qualifying_urls(url)
data <- scrape_qualifying_stats(quali_urls)
}
