% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_agr.R
\name{tab_agr}
\alias{tab_agr}
\title{Title}
\usage{
tab_agr(n, p)
}
\arguments{
\item{n}{année du GP}

\item{p}{numéro du GP}
}
\value{
a plot containing a representation of the evolution of the points during a season
}
\description{
Title
}
