% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dataset.R
\name{transf_data}
\alias{transf_data}
\title{Clean and transform the Qualification data
scraped from formula1.com.}
\usage{
transf_data(data)
}
\arguments{
\item{data}{Data.frame}
}
\value{

}
\description{
Clean and transform the Qualification data
scraped from formula1.com.
}
